+++ test/02-generate.js
@@ -640,13 +640,49 @@ describe('Module - generate', () => {
             height: 900
         }, assertCritical(target, expected, done));
     });
+
+    it('should handle PAGE_UNLOADED_DURING_EXECUTION error (inline)', done => {
+        const expected = read('fixtures/issue-314.html');
+        const target = '.issue-314.html';
+
+        critical.generate({
+            base: 'fixtures/',
+            src: 'issue-314.html',
+            css: ['fixtures/styles/bootstrap.css'],
+            dest: target,
+            inline: true,
+            width: 1300,
+            height: 900
+        }, assertCritical(target, expected, done));
+    });
+
+    it('should handle PAGE_UNLOADED_DURING_EXECUTION error', done => {
+        const expected = '';
+        const target = '.issue-314.css';
+
+        critical.generate({
+            base: 'fixtures/',
+            src: 'issue-314.html',
+            css: ['fixtures/styles/bootstrap.css'],
+            dest: target,
+            inline: false,
+            width: 1300,
+            height: 900
+        }, assertCritical(target, expected, done));
+    });
 });
 
 describe('Module - generate (remote)', () => {
     before(() => {
         const serve = serveStatic('fixtures', {index: ['index.html', 'index.htm']});
+        const serveUserAgent = serveStatic('fixtures/useragent', {
+            index: ['index.html', 'index.htm']
+        });
 
         this.server = http.createServer((req, res) => {
+            if (req.headers['user-agent'] === 'custom agent') {
+                return serveUserAgent(req, res, finalhandler(req, res));
+            }
             serve(req, res, finalhandler(req, res));
         });
 
@@ -1049,6 +1085,21 @@ describe('Module - generate (remote)', () => {
             dest: target,
             width: 1300,
             height: 900
+        }, assertCritical(target, expected, done));
+    });
+
+    it('should use the provided user agent to get the remote src', done => {
+        const expected = read('expected/generate-default.css');
+        const target = '.critical.css';
+
+        critical.generate({
+            base: 'fixtures/',
+            src: `http://localhost:${this.port}/generate-default-useragent.html`,
+            include: [/someRule/],
+            dest: target,
+            width: 1300,
+            height: 900,
+            userAgent: 'custom agent'
         }, assertCritical(target, expected, done));
     });
 });